requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.23.2
python-telegram-bot==22.5
gradio-client==1.13.3
requests==2.31.0
